move ChainWrapper to its own script
