+++ a2rchi/chains/chain_wrappers.py
@@ -1,0 +1,82 @@
+import os
+from typing import Any, Dict, List, Optional
+from langchain_core.language_models.base import BaseLanguageModel
+from langchain_core.prompts.base import BasePromptTemplate
+
+from a2rchi.chains.utils.callback_handlers import PromptLogger
+from a2rchi.chains.utils.token_limiter import TokenLimiter
+from a2rchi.utils.config_loader import load_config
+from a2rchi.utils.logging import get_logger
+
+logger = get_logger(__name__)
+
+# DEFINITIONS
+config = load_config()["chains"]["base"]
+data_path = load_config()["global"]["DATA_PATH"]
+
+class ChainWrapper:
+    """
+    Generic wrapper around Langchain's chains
+    to harmonize with our prompts and inputs.
+    """
+
+    def __init__(
+            self,
+            chain: Any,
+            llm: BaseLanguageModel,
+            prompt: BasePromptTemplate,
+            required_input_variables: List[str] = ['question'],
+            unprunable_input_variables: Optional[List[str]] = [],
+        ):
+        self.chain = chain
+        self.llm = llm
+        self.required_input_variables = required_input_variables
+        self.unprunable_input_variables = unprunable_input_variables
+        self.prompt = self._check_prompt(prompt)
+
+        self.prompt_logger = PromptLogger(os.path.join(data_path, config["logging"]["input_output_filename"]))
+        self.token_limiter = TokenLimiter(llm=self.llm, prompt=self.prompt)
+
+    def _check_prompt(self, prompt: BasePromptTemplate) -> BasePromptTemplate:
+        """
+        Check that the prompt is valid for this chain:
+            1. require that it contains all the required input variables
+        """
+        for var in self.required_input_variables:
+            if var not in prompt.input_variables:
+                raise ValueError(f"Chain requires input variable {var} in the prompt, but could not find it.")
+        return prompt
+    
+    def _prepare_payload(self, inputs: Dict[str, Any]) -> Dict[str, Any]:
+        """
+        Prepare the input_variables to be passed to the chain.
+        """
+
+        # grab all the input variables from the parameters given to the function
+        input_variables = {k:v for k,v in inputs.items() if k in self.prompt.input_variables}
+
+        # reduce number of tokens, if necessary
+        input_variables = self.token_limiter.prune_inputs_to_token_limit(**input_variables)
+
+        return input_variables
+
+    def invoke(self, inputs: Dict[str, Any]) -> Dict[str, Any]:
+        """
+        Call the chain to produce the LLM answer with some given inputs determined by the prompt.
+        """
+
+        # check if any of the unprunables are too large
+        for var in self.unprunable_input_variables:
+            if not self.token_limiter.check_input_size(inputs.get(var, "")):
+                return {"answer": self.token_limiter.INPUT_SIZE_WARNING.format(var=var)}
+
+        # get the payload
+        input_variables = self._prepare_payload(inputs)
+        
+        # produce LLM response
+        answer = self.chain.invoke(
+            input_variables,
+            config={"callbacks": [self.prompt_logger]}
+        )
+
+        return {"answer": answer, **input_variables}
